+++ src/Sql/Sql/ChangeLog.md
@@ -25,6 +25,7 @@
       `Get-AzSqlServerOutboundFirewallRule`
       `New-AzSqlServerOutboundFirewallRule`
       `Remove-AzSqlServerOutboundFirewallRule`
+* Fixed the identity logic for SystemAssigned,UserAssigned identities for New-AzSqlServer, New-AzSqlInstance
 * Updated cmdlets for getting and updating SQL database's differential backup frequency
       `Get-AzSqlDatabaseBackupShortTermRetentionPolicy`
       `Set-AzSqlDatabaseBackupShortTermRetentionPolicy`